Add files via upload
The streamlit app allows interactive investigation of the review data. The exploratory analysis section allows the user to look at map, charts and wordclouds at a town, category and detailed aspect level. The second section is a keyword analysis which allows the user to extract keywords, bigrams and noun phrases for each category split by positive and negative words. The final part of the app (not completed) is a sentiment analyser which allows the user to enter a single review and return the predicted aspects and sentiment based on the combined model developed in the report.